resume project with a cleaner folder

+++ src/main.cpp
@@ -20,11 +20,10 @@ color ray_color(const ray& r, const hittable_list& world, int depth) {
 }
 
 int main(){
-
     // Image
     const double aspect_ratio = 16.0 / 9.0;
     const int image_width = 400;
-    const int image_height = static_cast<int>(image_width / aspect_ratio);
+    const int image_height = (image_width / aspect_ratio);
     const int samples_per_pixel = 100; 
     const int max_depth = 50;
